Refactor: Split SchoolDetail.tsx into smaller components
Refactor SchoolDetail.tsx to improve maintainability by extracting sections into smaller, reusable components. This includes the header, overview, programs, and tuition sections.
